backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.22
great_expectations==0.18.22
pydantic-settings==2.7.0
python-multipart==0.0.20
# For AI
google-generativeai==0.8.3
openai==1.59.3
